twitter-project/twitterAPI/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2.12
djangorestframework==3.13.1
nest-asyncio==1.5.5
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
gunicorn==20.1.0
